requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==4.2.7
django-cors-headers==4.3.1
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
geographiclib==2.0
geopy==2.4.0
idna==3.10
Pillow==10.1.0
PyJWT==2.10.1
python-decouple==3.8
pytz==2025.2
requests==2.31.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
